Rename exptest.py to events.py to avoid conflict with function name
