+++ gammapy/time/__init__.py
@@ -2,6 +2,6 @@
 """Time analysis."""
 from .simulate import *
 from .lightcurve import *
-from .exptest import *
+from .events import *
 from .period import *
 from .plot_periodogram import *
