+++ gammapy/time/events.py
