+++ gammapy/time/tests/test_exptest.py
@@ -1,7 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from numpy.testing import assert_allclose
 from astropy.units import Quantity
-from ..exptest import exptest
+from ..events import exptest
 from ..simulate import random_times
